Altas a terminar (dom y clientes)

+++ database/migrations/2023_10_17_020428_create_clientes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('nombre');
             $table->string('email');
             $table->integer('dni');
-            $table->integer('telefono');
+            $table->bigInteger('telefono');
             $table->string('cuil');
             $table->timestamps();
         });
